@@ -5,63 +5,71 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "postinstall": "ngcc",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular-mdc/web": "^1.0.3",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.25",
-    "@angular/forms": "^8.2.14",
-    "@angular/http": "^7.1.1",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular-mdc/web": "^1.0.5",
+    "@angular/animations": "^9.1.13",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.13",
+    "@angular/compiler": "^9.1.13",
+    "@angular/core": "^9.1.13",
+    "@angular/forms": "9.1.13",
+    "@angular/localize": "^9.1.13",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "^9.1.13",
+    "@angular/platform-browser-dynamic": "^9.1.13",
+    "@angular/router": "^9.1.13",
     "@aspnet/signalr": "^1.1.4",
-    "@types/hammerjs": "^2.0.36",
-    "@types/simple-peer": "^9.6.0",
-    "angular-material": "^1.1.10",
+    "@ng-bootstrap/ng-bootstrap": "^6.2.0",
+    "@popperjs/core": "^2.11.6",
+    "@types/hammerjs": "^2.0.41",
+    "@types/simple-peer": "^9.11.5",
     "body-parser": "^1.18.3",
-    "config": "^3.2.2",
-    "core-js": "^2.5.4",
-    "express": "^4.16.3",
+    "bootstrap": "^5.2.1",
+    "config": "^3.3.8",
+    "core-js": "^2.6.12",
+    "express": "^4.18.1",
+    "file-saver": "^2.0.5",
     "hammerjs": "^2.0.8",
-    "material-components-web": "^2.2.0",
-    "ngx-image-cropper": "^2.0.2",
-    "node-sass": "^4.13.1",
+    "jquery": "^3.6.1",
+    "json-schema-ref-parser": "^9.0.9",
+    "material-components-web": "^2.3.1",
+    "ngx-image-cropper": "^2.1.2",
     "normalize.css": "^8.0.1",
     "package": "^1.0.1",
     "package.json": "^2.0.1",
-    "rxjs": "^6.5.2",
-    "simple-peer": "^9.6.2",
-    "zone.js": "~0.9.1"
+    "popper.js": "^1.16.1",
+    "rxjs": "^6.6.7",
+    "simple-peer": "^9.11.1",
+    "socket.io": "^4.5.2",
+    "socket.io-client": "^4.5.2",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "^0.901.15",
+    "@angular/cli": "^9.1.15",
+    "@angular/compiler-cli": "^9.1.13",
+    "@angular/language-service": "^9.1.13",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "@types/node": "~12.20.55",
+    "codelyzer": "^6.0.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.2.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.0",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.1",
+    "ng-swagger-gen": "^2.3.1",
+    "protractor": "^5.4.4",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.5.3"
+    "typescript": "~3.8.3"
   }
 }
