@@ -1,56 +1,58 @@
-import {Component, Inject, OnInit} from '@angular/core';
-import {FormControl} from '@angular/forms';
-import {Observable} from 'rxjs';
-import {Album} from '../../DataModules/album.model';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material';
-import {FileObject} from '../../DataModules/file.model';
-import {AlbumService} from '../../Services/album.service';
-import {map, startWith} from 'rxjs/operators';
+import { Component, Inject, OnInit } from '@angular/core';
+import { FormControl } from '@angular/forms';
+import { Observable } from 'rxjs';
+import { AlbumModel } from '../../DataModels/album.model';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { FileObject } from '../../DataModels/fileObject.model';
+import { AlbumService } from '../../Services/album.service';
+import { map, startWith } from 'rxjs/operators';
 
 @Component({
   selector: 'app-move-selected-dialog',
   templateUrl: './move-selected-dialog.component.html',
-  styleUrls: ['./move-selected-dialog.component.css']
+  styleUrls: ['./move-selected-dialog.component.css'],
 })
-export class MoveSelectedDialogComponent implements OnInit {
+export class MoveCopySelectedDialogComponent implements OnInit {
   myControl = new FormControl();
-  options?: Album[];
-  users: Observable<Album[]>;
-  something = '';
-  private selected;
+  options?: AlbumModel[];
+  albumObservable: Observable<AlbumModel[]>;
+  //something = '';
+  public selected;
 
-  constructor(private albumService: AlbumService,
-              public dialogRef: MatDialogRef<MoveSelectedDialogComponent>,
-              @Inject(MAT_DIALOG_DATA) public data: FileObject[]) {
-  }
+  constructor(
+    private albumService: AlbumService,
+    public dialogRef: MatDialogRef<MoveCopySelectedDialogComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: { files: FileObject[]; text: string }
+  ) {}
 
   ngOnInit() {
     this.albumService.getAlbums();
-    this.users = this.albumService.getAlbumUpdateListener();
-    this.users
-      .subscribe(
-        data => {
-          this.options = data;
-          // this.users2Show = data.sort(( l, r ) => {
-          //   const left = l.lastName + l.firstName;
-          //   const right = r.lastName + r.firstName;
-          //   return left < right ? -1 : 1;
-          // });
-        });
+    this.albumObservable = this.albumService.getAlbumUpdateListener();
+    this.albumObservable.subscribe((albums) => {
+      this.options = albums.filter((album)=>{return album.name!=='Recycle Bin'});
+      //
+      // this.users2Show = data.sort(( l, r ) => {
+      //   const left = l.LastName + l.FirstName;
+      //   const right = r.LastName + r.FirstName;
+      //   return left < right ? -1 : 1;
+      // });
+    });
 
-    this.users = this.myControl.valueChanges.pipe(
+    this.albumObservable = this.myControl.valueChanges.pipe(
       startWith(''),
-      map(value => this._filter(value))
+      map((value) => this._filter(value))
     );
   }
 
-  private _filter(value: string): Album[] {
+  private _filter(value: string): AlbumModel[] {
     const filterValue = value.toLowerCase();
 
-    return this.options.filter(option => option.AlbumName.toLowerCase().indexOf(filterValue) === 0);
+    return this.options.filter(
+      (option) => option.name.toLowerCase().indexOf(filterValue) === 0
+    );
   }
 
   onShopSelectionChanged(selected: any) {
-    this.selected = selected._id;
+    this.selected = selected; //.objId;
   }
 }
