@@ -1,13 +1,13 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {Subject} from 'rxjs';
-import {Album} from '../DataModules/album.model';
-import {ChatMessage} from '../DataModules/chat-message.model';
-import {map, tap} from 'rxjs/operators';
-import {host} from '../globals';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Subject } from 'rxjs';
+import { AlbumModel } from '../DataModels/album.model';
+import { ChatMessage } from '../DataModels/chat-message.model';
+import { map, tap } from 'rxjs/operators';
+import { host } from '../globals';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class ChatService {
   private messages: ChatMessage[] = []; // albums array
@@ -15,10 +15,8 @@ export class ChatService {
   private unreadMessageSize: number;
   private unreadMessageSizeUpdated = new Subject<number>();
   private messageUpdated = new Subject<ChatMessage[]>();
-  public currentlocalhost = 'https://localhost:5001';
 
-  constructor(private http: HttpClient) {
-  }
+  constructor(private http: HttpClient) {}
 
   /**
    * method responsible to get message observable
@@ -39,10 +37,14 @@ export class ChatService {
    * @param msg method responsible to send message
    */
   sendMessage(msg: ChatMessage) {
-    this.http.post(host + '/api/message/message/', msg)
-      .pipe(tap(console.log), map((serverResponse => serverResponse)))
+    this.http
+      .post(host + '/api/message/message/', msg)
+      .pipe(
+        tap(console.log),
+        map((serverResponse) => serverResponse)
+      )
       .subscribe((data) => {
-        console.log('Received Data:' + data);
+        console.log('sendMessage(msg: ChatMessage)');
         console.dir(data);
         this.messages.push(data);
         this.messageUpdated.next([...this.messages]);
@@ -54,28 +56,42 @@ export class ChatService {
    * @param receiverID method responsible to get all messages
    */
   getConversation(receiverID) {
-    this.http.get(host + '/api/message/messages/' + receiverID)
-      .pipe(tap(console.log), map(serverResponse => serverResponse)) // we re-edit the information to remove the  messages
-      .subscribe(files => {
-          this.messages = files;
-          console.log(JSON.stringify(this.messages));
+    console.log(`getConversation(${receiverID})`);
+    this.http
+      .get(host + '/api/message/messages/' + receiverID)
+      .pipe(
+        tap(console.log),
+        map((serverResponse) => serverResponse)
+      ) // we re-edit the information to remove the  messages
+      .subscribe((converasations) => {
+        console.dir(converasations);
+
+        if (converasations !== null) {
+          this.messages = converasations;
           this.messageUpdated.next(this.messages);
-          this.getUnreadMessageSize();// we return the information trough observable
+          this.getUnreadMessageSize(); // we return the information trough observable
+        } else {
+          this.messages = [];
+          this.messageUpdated.next(this.messages);
         }
-      );
+      });
   }
 
   /**
    * method responsible to get number of unread messages
    */
   getUnreadMessageSize() {
-    this.http.get(host + '/api/message/unread_messages/' + 'something')
-      .pipe(tap(console.log), map(serverResponse => serverResponse)) // we re-edit the information to remove the  messages
-      .subscribe(files => {
-          this.unreadMessageSize = files;
-          console.log(JSON.stringify(this.unreadMessageSize));
-          this.unreadMessageSizeUpdated.next(this.unreadMessageSize); // we return the information trough observable
-        }
-      );
+    this.http
+      .get(host + '/api/message/unread_messages/' + 'something')
+      .pipe(
+        tap(console.log),
+        map((serverResponse) => serverResponse)
+      ) // we re-edit the information to remove the  messages
+      .subscribe((files) => {
+        console.log('getUnreadMessageSize()');
+        this.unreadMessageSize = files;
+        console.log(JSON.stringify(this.unreadMessageSize));
+        this.unreadMessageSizeUpdated.next(this.unreadMessageSize); // we return the information trough observable
+      });
   }
 }
