@@ -1,29 +1,30 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {map, tap} from 'rxjs/operators';
-import {BehaviorSubject, Observable} from 'rxjs';
-import {FileObject} from '../DataModules/file.model';
-import {Album} from '../DataModules/album.model';
-import {host} from '../globals';
-import {forEachComment} from 'tslint';
+import { Injectable } from "@angular/core";
+import { HttpClient } from "@angular/common/http";
+import { map, tap } from "rxjs/operators";
+import { BehaviorSubject, Observable } from "rxjs";
+import { FileObject } from "../DataModels/fileObject.model";
+//import { AlbumModel } from '../DataModels/album.model';
+import { host } from "../globals";
+import { NewAlbumDialogComponent } from "../Dialogs/new-album-dialog/new-album-dialog.component";
+import { AlbumModel } from "../DataModels/album.model";
+//import { forEachComment } from 'tslint';
 
 // we declare that this service should be created
 // by the root application injector.
-@Injectable({providedIn: 'root'})
-
+@Injectable({ providedIn: "root" })
 export class FileService {
   private files: FileObject[] = []; // files array of type FileObject (interface)
   private sortedFiles: FileObject[] = [];
-  private filesTodelete: string[] = [];
-  private filesUpdated = new BehaviorSubject<FileObject[]>(this.files); // filesUpdated is the subject with BehaviorSubject returns files
-  public currentlocalhost = 'https://localhost:5001';
+  //private filesTodelete: string[] = [];
+  private filesUpdatedList = new BehaviorSubject<FileObject[]>(this.files); // filesUpdated is the subject with BehaviorSubject returns files
+  // Should be in config
+  //public currentlocalhost = "https://localhost:5001";
 
   /**
    * constructor that inits Http client protocol
    * @param {HttpClient} http
    */
-  constructor(private http: HttpClient) {
-  }
+  constructor(private http: HttpClient) {}
 
   /**
    * we return filesUpdated asObservable trough
@@ -31,49 +32,64 @@ export class FileService {
    * @returns {Observable<FileObject[]>} as data that couldn't not be changed
    */
   getFileUpdateListener(): Observable<FileObject[]> {
-    return this.filesUpdated.asObservable();
+    return this.filesUpdatedList.asObservable();
   }
 
   clearFiles() {
-    this.filesUpdated.next(null);
+    this.filesUpdatedList.next(null);
   }
 
   /**
    * method responsible to upload files to the server
-   * @param {string} albumName - the album to link the files to
+   *
    * @param {FormData} formData - form data information
    */
-  uploadFile(albumId: string, formData: FormData) {
-    formData.append('AlbumId', albumId); // we append the album name to the file
+  uploadFile(/* albumobjId: string, */ formData: FormData) {
+    //formData.append("AlbumId", albumId); // we append the album name to the file
     console.log(formData); // log for debug will be removed in the future
-    const request$ = this.http.post(host + '/api/data/file', formData, {reportProgress: true})
+
+    const request$ = this.http
+      .post(host + "/api/Data/uploadfile", formData, { reportProgress: true })
       // the server returns the file object that was entered the db
-      .pipe(tap(console.log), map((serverResponse => serverResponse.file)))
-      .subscribe((file) => {
-        console.log('Received Data:');
-        const index = this.files.findIndex((e) => e.url === file.url);
-        if (index === -1) {
-          this.files.push(file);
-        } else {
-          this.files[index] = file;
+      .pipe(
+        tap(console.log),
+        map((serverResponse) => serverResponse.file)
+      )
+      .subscribe(
+        (file) => {
+          console.log("Received Data:");
+          const index = this.files.findIndex((e) => e.url === file.url);
+          if (index === -1) {
+            this.files.push(file);
+          } else {
+            this.files[index] = file;
+          }
+          // this.files.push(file); // we push the received file to the fleUpdated list
+          this.filesUpdatedList.next([...this.files]); // and notification to the Subject to return the new files list.
+        },
+        (error) => {
+          console.log(error);
         }
-        //this.files.push(file); // we push the received file to the fleUpdated list
-        this.filesUpdated.next([...this.files]); // snd notification to the Subject to return the new files list.
-      });
+      );
   }
 
   uploadAvatar(formData: FormData) {
-    console.log(formData); // log for debug will be removed in the future
-    return this.http.post<{ url: string }>(host + '/api/data/avatar', formData, {reportProgress: true});
-    // the server returns the file object that was entered the db
-    /*      .subscribe((file) => {
-            console.dir(file);
-          });*/
+    // console.log(formData); // log for debug will be removed in the future
+    return this.http.post<{ url: string }>(
+      host + "/api/data/avatar",
+      formData,
+      { reportProgress: true }
+    );
   }
 
+  /**
+   * this method pushes a FileObject to list of Files
+   * and to filesUpdated list
+   * @param tmp
+   */
   insertFile(tmp: FileObject) {
     this.files.push(tmp);
-    this.filesUpdated.next([...this.files]);
+    this.filesUpdatedList.next([...this.files]);
   }
 
   /**
@@ -83,19 +99,53 @@ export class FileService {
    * @returns {Subscription}
    */
   getFiles(albumId: string) {
-    return this.http.get<{ message: string, files: any }>(host + '/api/data/file/' + albumId)
-      .pipe(tap(console.log), map(serverResponse => serverResponse.files)) // we re-edit the information to remove the  messages
-      .subscribe(files => {
-          this.files = files;
-          console.log(JSON.stringify(this.files));
-          this.filesUpdated.next(this.files); // we return the information trough observable
-        }
-      );
+    const getFileSubs = this.http
+      .get<{ message: string; files: any }>(
+        host + "/api/data/getAlbumFiles/" + albumId
+      )
+      .pipe(
+        tap(console.log),
+        map((serverResponse) => serverResponse.files)
+      ); // we re-edit the information to remove the  messages
+
+    return getFileSubs.subscribe((files) => {
+      this.files = files;
+      //console.log(JSON.stringify(this.files));
+      //console.dir(files);
+      this.filesUpdatedList.next(this.files); // we return the information trough observable
+    });
   }
 
-  updateLocalFiles(fileObject: FileObject) {
-    this.files = this.files.filter(obj => obj !== fileObject);
-    this.filesUpdated.next(this.files);
+  /**
+   * NEW - OFF
+   * method responsible to retrieve files list from the server
+   * by album id
+   * @param {string} albumId
+   * @returns {Subscription}
+   */
+  /* getFiles2(albumFilesId: string[]) {
+    console.warn('Sending Rquest');
+    const getFileSubs = this.http
+    .post<{ message: string; files: any }>(host + '/api/data/files/', albumFilesId)
+    .pipe(
+      tap(console.log),
+      map((serverResponse) => serverResponse.files)
+    ); // we re-edit the information to remove the  messages
+    
+    return getFileSubs.subscribe((files) => {
+        this.files = files;
+        console.log(JSON.stringify(this.files));
+        this.filesUpdatedList.next(this.files); // we return the information trough observable
+      });
+  }
+ */
+  /**
+   *
+   * @param fileObject This method filters out failed upload files?
+   */
+  filterOutLocalFilesList(fileObject: FileObject) {
+    this.files = this.files.filter((obj) => obj !== fileObject);
+    this.filesUpdatedList.next(this.files);
   }
 
   getCurrentFiles() {
@@ -105,37 +155,53 @@ export class FileService {
     return this.files;
   }
 
-  /**
-   * mrthod responsibel to delete file from the server
-   * @param {string} file_id
-   */
-  delFileFromDb(file_id: string, album_id: string) {
-    const fileToDelete = {FileId: file_id, Album: album_id};
-    this.http.post(host + '/api/data/delete_file/', fileToDelete)
+  delFromAllUserDb(
+    listOfFilesToDelete: string[],
+    looseOwnership: boolean = false
+  ) {
+    const fileToDelete = {
+      ListOfFilesToDelete: listOfFilesToDelete,
+      Album: null,
+      LooseOwnership: looseOwnership,
+    };
+    this.http
+      .post(host + "/api/data/deleteFileFromAllMyAlbums/", fileToDelete)
       .subscribe(() => {
-        console.log('deleted!');
-        const idx = this.files.findIndex(val => val._id === file_id); // we search for te file index inside the file array
-        this.files.splice(idx, 1); // remove it from the list
-        this.filesUpdated.next([...this.files]); // return new list as observable
+        console.log("delFromAllUserDb - deleted!");
+
+        listOfFilesToDelete.forEach((fileId) => {
+          this.files.splice(
+            this.files.findIndex((val) => val.objId === fileId),
+            1
+          );
+        });
+        this.filesUpdatedList.next([...this.files]); // return new list as observable
       });
   }
 
-  delFromAllUserDb(file_id: string) {
-    const fileToDelete = {FileId: file_id, Album: null};
-    this.http.post(host + '/api/data/delete_file_from_all/', fileToDelete)
+  shareFileWithUser(currentAlbum: string, userId: string, files: string[]) {
+    const fileToShare = {
+      SourceAlbumId: currentAlbum,
+      UserToShareWithObjId: userId,
+      ListFileIds: files,
+    };
+    this.http
+      .post(host + "/api/data/share_selected/", fileToShare)
       .subscribe(() => {
-        console.log('deleted!');
-        const idx = this.files.findIndex(val => val._id === file_id); // we search for te file index inside the file array
-        this.files.splice(idx, 1); // remove it from the list
-        this.filesUpdated.next([...this.files]); // return new list as observable
+        console.log("files Shared!");
       });
   }
 
-  shareFileWithUser(user_id: string, files: string[]) {
-    const fileToShare = {UserId: user_id, Files: files};
-    this.http.post(host + '/api/data/share_selected/', fileToShare)
+  setOwnership(userId: string, newAlbum: AlbumModel, listOfFiles: string[]) {
+    const ownership = {
+      NewOwnerId: userId,
+      Album: newAlbum,
+      ListFileIds: listOfFiles,
+    };
+    this.http
+      .post(host + "/api/data/setOwner_selected/", ownership)
       .subscribe(() => {
-        console.log('files Shared!');
+        console.log("setOwnership - done!");
       });
   }
 
@@ -143,21 +209,36 @@ export class FileService {
    * method responsible to delete file by id
    * @param id
    */
-  addRemoveDelete(id: string) {
-    this.filesTodelete.push(id);
+  /*   addRemoveDelete(objId: string) {
+    this.filesTodelete.push(objId);
     console.log(this.filesTodelete);
   }
 
+ */
   filterFilesByDate(startDate, endDate) {
-
-    const dataNew = this.files.filter(product => product.exifData.some(item => item.name === 'Date/Time Original' &&
-      (new Date(item.value.split(' ')[0].replace(':', '-'))) >= (new Date(startDate)) &&
-      (new Date(item.value.split(' ')[0].replace(':', '-'))) <= (new Date(endDate))));
-    this.filesUpdated.next([...dataNew]);
+    const dataNew = this.files.filter((product) => {
+      new Date(
+        product.exifData.dateTimeOriginal
+          .toString()
+          .split(" ")[0]
+          .replace(":", "-")
+      ) >= new Date(startDate) &&
+        new Date(
+          product.exifData.dateTimeOriginal
+            .toString()
+            .split(" ")[0]
+            .replace(":", "-")
+        ) <= new Date(endDate);
+    });
+    this.filesUpdatedList.next([...dataNew]);
     this.sortedFiles = dataNew;
   }
 
   resetFilter() {
-    this.filesUpdated.next([...this.files]);
+    this.filesUpdatedList.next([...this.files]);
+  }
+
+  ngOnDestroy() {
+    this.filesUpdatedList.unsubscribe();
   }
 }
