@@ -1,6 +1,8 @@
 import {Component, EventEmitter, OnDestroy, OnInit, Output} from '@angular/core';
 import {AuthenticationService} from '../../Services/auth.service';
 import {observable, Subscription} from 'rxjs';
+import { OptionButtonsService } from 'src/app/Services/tool-bar-buttons.service';
+import { SideNavLink } from 'src/app/DataModels/side-nav-link.model';
 
 @Component({
   selector: 'app-sidenav',
@@ -10,11 +12,30 @@ import {observable, Subscription} from 'rxjs';
 export class SidenavComponent implements OnInit, OnDestroy {
   userIsAuthenticated = JSON.parse(localStorage.getItem('loggedIn') || 'false');
   userName: string;
+
+/* 
+  sideNavElements: SideNavLink[] = [
+    {
+      displayName: 'Posts',
+      routLink: '//post',
+      iconName: 'announcement',
+    },
+    {
+      displayName: 'Groups',
+      iconName: 'announcement',
+      routLink: '/groups',
+    },
+
+  ]; */
+
+
+
   avatar = '/assets/images/man.svg';
-  private authListenerSubs: Subscription;
+  public authListenerSubs: Subscription;
+
   @Output() closeSidenav = new EventEmitter<void>();
 
-  constructor(private authService: AuthenticationService) {
+  constructor(private authService: AuthenticationService, private optionButtonsService: OptionButtonsService) {
   }
 
   /**
@@ -24,13 +45,13 @@ export class SidenavComponent implements OnInit, OnDestroy {
   ngOnInit() {
     this.authListenerSubs = this.authService.getAuthStatusListener().subscribe(isAuthenticated => {
       this.userIsAuthenticated = isAuthenticated;
-      if(this.userIsAuthenticated===true){
+      if (this.userIsAuthenticated === true) {
         this.userName = this.authService.getUserName();
         this.avatar = this.authService.getAvatar();
       }
 
     });
-    if(this.authService.getAuthStatus() === true){
+    if (this.authService.getAuthStatus() === true) {
       this.authService.updateAuthStatusListener();
     }
   }
@@ -46,8 +67,10 @@ export class SidenavComponent implements OnInit, OnDestroy {
   /**
    * function responsible to emit event of the closing of side bar
    */
-  onClose() {
+  onClose(buttonsSetName: string) {
+    console.log(` onClose(buttonsSetName: string)`);
+    this.optionButtonsService.setButtons(buttonsSetName);
+    //this.toolBarButtonsService.onViewUpdatedEventEmiter.emit(buttonsSetName);
     this.closeSidenav.emit();
   }
-
 }
