@@ -2,7 +2,7 @@
 import {AuthenticationService} from '../Services/auth.service';
 import {Injectable} from '@angular/core';
 import {CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router, UrlTree} from '@angular/router';
-import {Observable} from 'rxjs/internal/Observable';
+import {Observable} from 'rxjs';
 
 /**
  * its injectable components AuthGuard that implements CanActive
@@ -15,8 +15,9 @@ export class AuthGuard implements CanActivate {
    * @param {AuthenticationService} authService
    * @param {Router} router
    */
-  constructor(private authService: AuthenticationService, private router: Router) {
-  }
+  constructor(
+    private authService: AuthenticationService,
+    private router: Router) {/* Default Empty */ }
 
   /**
    * canActivate method checks if user is authenticated before allowing router options
@@ -24,25 +25,13 @@ export class AuthGuard implements CanActivate {
    * @param {RouterStateSnapshot} state
    * @returns {boolean | Observable<boolean> | Promise<boolean>}
    */
-
   canActivate(
     rout: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): boolean | Observable<boolean> | Promise<boolean> {
     const isAuth = this.authService.getAuthStatus(); // gets authentication status from service
     if (!isAuth) {
-      this.router.navigate(['/']); // if not authenticated moves user back to login menu
+      this.router.navigate(['/register/login']); // if not authenticated moves user back to login menu
     }
     return isAuth; // return the status
   }
-
-
-  /*  canActivate(route: ActivatedRouteSnapshot,
-                state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-      const isAuth = this.authService.getAuthStatus(); // gets authentication status from service
-      if (!isAuth) {
-        this.router.navigate(['/']); // if not authenticated moves user back to login menu
-      }
-      return isAuth; // return the status
-    }*/
-
 }
