@@ -2,7 +2,7 @@ import { Component, OnInit, AfterViewChecked, ViewChild, ElementRef } from '@ang
 import {ChatService} from '../../Services/chat.service';
 import {ActivatedRoute, ParamMap} from '@angular/router';
 import {Observable} from 'rxjs';
-import {ChatMessage} from '../../DataModules/chat-message.model';
+import {ChatMessage} from '../../DataModels/chat-message.model';
 
 @Component({
   selector: 'app-message-feed',
@@ -14,13 +14,19 @@ export class MessageFeedComponent implements OnInit{
   messages: Observable<ChatMessage[]>;
   @ViewChild('scroller') private feedContainer: ElementRef;
 
-  constructor(private route: ActivatedRoute, private chatService: ChatService) {
+  constructor(
+    private route: ActivatedRoute,
+    private chatService: ChatService) {
   }
-
+  
   ngOnInit() {
     this.route.paramMap.subscribe((params: ParamMap) => {
-      console.log(params.get('id'));
+      console.log(`MessageFeedComponent.ngOnInit`);
+      
+      //this.receiver = this.authService.getUserID();
       this.receiver = params.get('id');
+      console.warn(`This receiver is: ${this.receiver}`);
+
       this.chatService.getConversation(this.receiver);
       this.messages = this.chatService.getMessagetUpdateListener();
     });
