@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import * as signalR from "@aspnet/signalr";
 import { Subject } from 'rxjs';
-import {SignalInfo, UserInfo} from '../DataModules/video.model';
+import {SignalInfo, UserInfo} from '../DataModels/video.model';
 import {host} from '../globals';
 
 
@@ -28,37 +28,50 @@ export class VideoService {
 
   public async startConnection(currentUser: string): Promise<void> {
 
+    // ths is the HUB connection setting
     this.hubConnection = new signalR.HubConnectionBuilder()
-      .withUrl(host +'/signalrtc')
+      .withUrl(host + '/videostream')
       .build();
 
+    // this is starting connection to HUB
     await this.hubConnection.start();
     console.log('Connection started');
 
-    this.hubConnection.on('NewUserArrived', (data) => {
-      this.newPeer.next(JSON.parse(data));
+    // registration of 'NewUserArrived' method
+    this.hubConnection.on('NewUserArrived', (username) => {
+      this.newPeer.next(JSON.parse(username));
     });
 
+    // registration of 'UserSaidHello' method
     this.hubConnection.on('UserSaidHello', (data) => {
       this.helloAnswer.next(JSON.parse(data));
     });
 
+    // registration of 'UserDisconnect' method
     this.hubConnection.on('UserDisconnect', (data) => {
       this.disconnectedPeer.next(JSON.parse(data));
     });
 
+    // registration of 'SendSignal' method
     this.hubConnection.on('SendSignal', (user, signal) => {
-      this.signal.next({ user, signal });
+      this.signal.next({ user: user, signal: signal });
     });
 
+    // this is calling - invoking the request to HUB server
+    // and calling 'NewUser' methodthat accepts currentUser
+    // as argument
     this.hubConnection.invoke('NewUser', currentUser);
   }
 
+  // public method to be called in angular
   public sendSignalToUser(signal: string, user: string) {
+    // invoking 'SendSignal' method on HUB
     this.hubConnection.invoke('SendSignal', signal, user);
   }
 
+  // public method to be called in angular
   public sayHello(userName: string, user: string): void {
+    // invoking 'HelloUser' method on HUB
     this.hubConnection.invoke('HelloUser', userName, user);
   }
 
